backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pytest==7.4.3

# Optional: Only needed if using PostgreSQL (not required for SQLite)
# Uncomment if you want to use PostgreSQL instead of SQLite
# psycopg2-binary==2.9.9
